participants loaded into participants provider in conversation list activity by placing conversation list initialization in parse callback function
